+++ spec/support/formatter_support.rb
@@ -32,13 +32,25 @@ module FormatterSupport
   if RUBY_VERSION.to_f < 1.9
     def expected_summary_output_for_example_specs
       <<-EOS.gsub(/^\s+\|/, '').chomp
-        |Pending:
-        |  pending spec with no implementation is pending
-        |    # Not yet implemented
-        |    # ./spec/rspec/core/resources/formatter_specs.rb:4
-        |  pending command with block format with content that would fail is pending
-        |    # No reason given
-        |    # ./spec/rspec/core/resources/formatter_specs.rb:9
+        |Pending: (Failures listed here are expected and do not affect your suite's status)
+        |
+        |  1) pending spec with no implementation is pending
+        |     # Not yet implemented
+        |     # ./spec/rspec/core/resources/formatter_specs.rb:4
+        |
+        |  2) pending command with block format with content that would fail is pending
+        |     # No reason given
+        |     Failure/Error: expect(1).to eq(2)
+        |
+        |       expected: 2
+        |            got: 1
+        |
+        |       (compared using ==)
+        |     # ./spec/rspec/core/resources/formatter_specs.rb:11
+        |     # ./spec/support/formatter_support.rb:13:in `run_example_specs_with_formatter'
+        |     # ./spec/support/sandboxing.rb:16
+        |     # ./spec/support/sandboxing.rb:14
+        |     # ./spec/support/sandboxing.rb:8
         |
         |Failures:
         |
@@ -85,13 +97,25 @@ module FormatterSupport
   else
     def expected_summary_output_for_example_specs
       <<-EOS.gsub(/^\s+\|/, '').chomp
-        |Pending:
-        |  pending spec with no implementation is pending
-        |    # Not yet implemented
-        |    # ./spec/rspec/core/resources/formatter_specs.rb:4
-        |  pending command with block format with content that would fail is pending
-        |    # No reason given
-        |    # ./spec/rspec/core/resources/formatter_specs.rb:9
+        |Pending: (Failures listed here are expected and do not affect your suite's status)
+        |
+        |  1) pending spec with no implementation is pending
+        |     # Not yet implemented
+        |     # ./spec/rspec/core/resources/formatter_specs.rb:4
+        |
+        |  2) pending command with block format with content that would fail is pending
+        |     # No reason given
+        |     Failure/Error: expect(1).to eq(2)
+        |
+        |       expected: 2
+        |            got: 1
+        |
+        |       (compared using ==)
+        |     # ./spec/rspec/core/resources/formatter_specs.rb:11:in `block (3 levels) in <top (required)>'
+        |     # ./spec/support/formatter_support.rb:13:in `run_example_specs_with_formatter'
+        |     # ./spec/support/sandboxing.rb:16:in `block (4 levels) in <top (required)>'
+        |     # ./spec/support/sandboxing.rb:14:in `block (3 levels) in <top (required)>'
+        |     # ./spec/support/sandboxing.rb:8:in `block (2 levels) in <top (required)>'
         |
         |Failures:
         |