+++ Changelog.md
@@ -38,6 +38,8 @@ Enhancements:
   `RSpec.world` state. (Tyler Ball, 1808)
 * Apply line-number filters only to the files they are scoped to,
   allowing you to mix filtered and unfiltered files. (Myron Marston, #1839)
+* When dumping pending examples, include the failure details so that you
+  don't have to un-pend the example to see it. (Myron Marston, #1844)
 
 Bug Fixes: